correción con tsx y utilizando componente padre e hijo
