+++ src/TaskManager.tsx
@@ -1,0 +1,37 @@
+import React, { useState } from "react";
+import TaskForm from "./TaskForm";
+import TaskList from "./TaskList";
+
+interface Task {
+  text: string;
+  completed: boolean; 
+}
+
+function TaskManager() {
+  const [tasks, setTasks] = useState<Task[]>([]);
+
+  const handleCompleted = (index: number): void => {
+    const updatedTasks = [...tasks];
+
+    updatedTasks[index].completed = !updatedTasks[index].completed;
+
+    setTasks(updatedTasks);
+  }
+
+  function createTask(taskText: string) {
+    const newTask: Task = { text: taskText, completed: false };
+
+    setTasks([...tasks, newTask]);
+  }
+
+  return (
+    <main>
+      <div className="container-wrapper main-content">
+        <TaskForm createTask={createTask} />
+        <TaskList tasks={tasks} handleCompleted={handleCompleted} />
+      </div>
+    </main>
+  );
+}
+
+export default TaskManager;
