--- src/tasks.js
@@ -1,3 +1,0 @@
-export const tasks = [
-    {text: "hola"}
-];
